Moved sqlite test out of DbSessionTest. Renamed DbSessionTest to AbstractDbSessionTest
Fixed tests to work corrctly even in case previous test did not clean up migrations table
