+++ tests/framework/web/session/sqlite/DbSessionTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace yiiunit\framework\web\session\mysql;
+namespace yiiunit\framework\web\session\sqlite;
 
 /**
  * Class DbSessionTest
@@ -8,12 +8,12 @@ namespace yiiunit\framework\web\session\mysql;
  * @author Dmytro Naumenko <d.naumenko.a@gmail.com>
  *
  * @group db
- * @group mysql
+ * @group sqlite
  */
 class DbSessionTest extends \yiiunit\framework\web\session\AbstractDbSessionTest
 {
     protected function getDriverNames()
     {
-        return ['mysql'];
+        return ['sqlite'];
     }
 }
